add : nutritionhistory entity and REST API

See merge request s06-ai-image-sub2/S06P22E102!102

Former-commit-id: 8a78419fe07535612e0a46c918b69bdb5426cf89

+++ back/src/main/java/com/aisher/helf/db/entity/Food.java
@@ -26,13 +26,13 @@ public class Food {
     int kcal;
 
     @Column(name = "carbohydrate", scale = 2)
-    BigDecimal carbohydrate;
+    double carbohydrate;
 
     @Column(name = "protein", scale = 2)
-    BigDecimal protein;
+    double protein;
 
     @Column(name = "fat", scale = 2)
-    BigDecimal fat;
+    double fat;
 
     public void updateFood(FoodRegisterReq foodRegisterReq) {
         this.foodName = foodRegisterReq.getFoodName();
